// packages/react/src/components/Accordion/Root/types.ts
import type { ReactNode } from 'react';

export type AccordionRootProps =
  | {
      children?: ReactNode;
      type?: 'single';
      value?: string;
      defaultValue?: string;
      onValueChange?: (value: string) => void;
      collapsible?: boolean;
      disabled?: boolean;
    }
  | {
      children?: ReactNode;
      type: 'multiple';
      value?: string[];
      defaultValue?: string[];
      onValueChange?: (value: string[]) => void;
      collapsible?: never;
      disabled?: boolean;
    };